requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Session==0.4.0
pymysql==1.0.2
